Edit tile class to make rect collision detection easier, implemented some helpful code to showcase how collision will work & mouse visuals

+++ src/main/java/org/goose/objects/Player.java
@@ -60,7 +60,7 @@ public class Player extends Entity{
         }
 
         //do gravity
-        this.getVelocity().y += Main.world.getGravity()/20;
+        this.getVelocity().y += (float) (Main.world.getGravity()/20);
 
         //Add the rect position with the velocity
         this.getRect().x += this.getVelocity().x;
